Upload final version of the website (with minor errors)
I cant figure out what "left" is needed for in the css, or how to implement bold font in the html without adding a style tag (as adding it as a class didn't work?). This was running properly on all screen sizes, but I'm not sure why the page height is now showing as wrong on markbot. Markbot was then showing my page differently than it loads in my browser after whatever happened here, so it's become mostly impossible to figure out what it wants. Should have done more commits to keep a copy of it when markbot was accepting it still, sorry.